Don't use blobs
Safari throws a UnknownError when trying to save a Blob to IndexedDB, so we needed a new way to store shit.

Native ArrayBuffer and TypedArray serialization are slow as dogshit, so I wrote my own thing to convert the audio data into giant utf-16 strings.

When possible, we use the new TextDecoder class that WHATWG has been working on. If TextDecoder isn't present, we do that shit manually with some batched calls to String.fromCharCode (batched to prevent too many string concatenations, but not batched so much that we blow up the call stack by trying to apply() a million arguments).

Miraculousy, this is all pretty close to the performance of using blobs, even when TextDecoder isn't available.

Fucking JavaScripts, man. What can't they do?